business-logic/src/main/java/parser/CalculationObject.java
package parser;

public class CalculationObject {
	private String operand1;
	private String operator ;
	private String operand2;
	
	public CalculationObject() {
	}

	public CalculationObject(String operand1, String operator, String operand2) {
		this.operand1 = operand1;
		this.operator = operator;
		this.operand2 = operand2;
	}

	public String getOperand1() {
		return operand1;
	}

	public void setOperand1(String operand1) {
		this.operand1 = operand1;
	} 

	public String getOperator() {
		return operator;
	}

	public void setOperator(String operator) {
		this.operator = operator;
	}

	public String getOperand2() {
		return operand2;
	}

	public void setOperand2(String operand2) {
		this.operand2 = operand2;
	}
}